+++ canvas-component-tools/CHANGELOG.md
@@ -7,6 +7,7 @@ Canvas Component Tools uses [semantic versioning](https://semver.org/).
 * Added theme blocks support
 * Added `--debug` flag to remove restrictions
 * Updated order of questions and templates
+* Updated to prevent component handle parts starting with a number
 * Updated to not error if empty component folder already exists
 * Updated internal folder structure
 * Fixed formatting issue when creating multiple web components ([CANVAS-608](https://we-make-websites.atlassian.net/browse/CANVAS-608))